# specs/001-farm-ops-platform/contracts/api.yaml
openapi: 3.1.0
info:
  title: FERM+ API
  version: 0.1.0
  description: >
    Contract for the initial FERM+ web platform covering authentication,
    farm management, intelligent agenda, alerts, and core operational modules.
servers:
  - url: https://api.ferm-plus.local
...
paths:
  /auth/session:
    post:
      tags: [Auth]
      summary: Sign in and create a user session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
...
  /farms/{farmId}:
    get:
      tags: [Farms]
      summary: Get detailed farm workspace information
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: Farm detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmDetail'
    patch:
      tags: [Farms]
      summary: Update farm information
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: Farm updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmDetail'
  /farms/{farmId}/agenda:
    get:
      tags: [Agenda]
      summary: Get agenda data for a farm
      parameters:
        - $ref: '#/components/parameters/FarmId'
        - in: query
          name: view
          schema:
            type: string
            enum: [daily, weekly, monthly, list]
      responses:
        '200':
          description: Agenda payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgendaResponse'
  /farms/{farmId}/agenda/tasks:
    post:
      tags: [Agenda]
      summary: Create or regenerate a task set for a farm
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '201':
          description: Task set created
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgendaTask'
  /farms/{farmId}/alerts:
    get:
      tags: [Alerts]
      summary: List alerts for a farm
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
...
  /farms/{farmId}/crops/plots:
    get:
      tags: [Crops]
      summary: List crop plots
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: Crop plot list
    post:
      tags: [Crops]
      summary: Create a crop plot
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '201':
          description: Crop plot created
...
components:
  parameters:
    FarmId:
      in: path
      name: farmId
      required: true
      schema:
        type: string
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SessionResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/CurrentUser'
    CurrentUser:
      type: object
      required: [id, fullName, email, role]
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, PROPRIETAIRE]
        assignedFarmIds:
          type: array
          items:
            type: string
    FarmSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        activityType:
          type: string
        location:
          type: string
    FarmDetail:
      allOf:
        - $ref: '#/components/schemas/FarmSummary'
        - type: object
          properties:
            description:
              type: string
            surfaceArea:
              type: number
            ownerUserId:
              type: string
              nullable: true
    AgendaResponse:
      type: object
      properties:
        today:
          type: array
          items:
            $ref: '#/components/schemas/AgendaTask'
        upcoming:
          type: array
          items:
            $ref: '#/components/schemas/AgendaTask'
        completed:
          type: array
          items:
            $ref: '#/components/schemas/AgendaTask'
        overdue:
          type: array
          items:
            $ref: '#/components/schemas/AgendaTask'
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
    AgendaTask:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
        priority:
          type: string
        status:
          type: string
          enum: [A_FAIRE, EN_COURS, TERMINEE, EN_RETARD]
        isAutoGenerated:
          type: boolean
    Alert:
      type: object
      properties:
        id:
          type: string
        alertType:
          type: string
        severity:
          type: string
        title:
          type: string
        message:
          type: string
        triggeredAt:
          type: string
          format: date-time